[android] move egl under renderer package
